feat(backend): refactor and document
This commit significantly cleans up the routes and middleware stack, and
adds documentation to the same.

The refactor resolves several problems with routes where format
transformation, authentication, and others were not happening in the
right order. It also removes several redundant or unnecessary
middlewares.

The result is a working Transit API + server-side rendered HTML pages.