feat(entity): do not rely on SQL cascading
I’m spending way too much time on this!

Previously we were setup so that when a system is deleted, the database
would also automatically cascade the deletion to the system’s parts and
relationships, however that is not NERDY ENOUGH for me, because what if
we want to add some BUSINESS LOGIC when each part is deleted.

So now we delete manually, like god intended.